feat: remove fullpage
