+++ frontend/src/components/main/LastView.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+import Footer from "../Footer";
+export default function LastView() {
+  return (
+    <div>
+      <Footer />
+    </div>
+  );
+}
